Make file tests into integration tests
