+++ src/matchers/existing_path.rs
@@ -1,6 +1,7 @@
 // Copyright 2014 Carl Lerche, Steve Klabnik, Alex Crichton, Ben Longbons,
 //                Michael Gehring, Yehuda Katz
 // Copyright 2015 Carl Lerche, Alex Crichton, Graham Dennis, Robin Gloster
+// Copyright 2016 Urban Hafner
 //
 // Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
 // http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
@@ -75,43 +76,4 @@ pub fn existing_file() -> ExistingPath {
 
 pub fn existing_dir() -> ExistingPath {
     ExistingPath { path_type: PathType::Dir }
-}
-
-#[cfg(test)]
-mod test {
-    use std::env;
-    use std::borrow::ToOwned;
-    use std::path::{Path, PathBuf};
-    use {assert_that, is, is_not, existing_file, existing_dir, existing_path};
-
-    #[test]
-    fn an_existing_file() {
-        let path = path(env::var("TEST_EXISTS_FILE"), "./README.md");
-
-        assert_that(&path, is(existing_path()));
-        assert_that(&path, is(existing_file()));
-        assert_that(&path, is_not(existing_dir()));
-    }
-
-    #[test]
-    fn an_existing_dir() {
-        let path = path(env::var("TEST_EXISTS_DIR"), "./target");
-
-        assert_that(&path, is(existing_path()));
-        assert_that(&path, is(existing_dir()));
-        assert_that(&path, is_not(existing_file()));
-    }
-
-    #[test]
-    fn a_nonexisting_path() {
-        let path = path(env::var("TEST_EXISTS_NONE"), "./zomg.txt");
-
-        assert_that(&path, is_not(existing_path()));
-        assert_that(&path, is_not(existing_file()));
-        assert_that(&path, is_not(existing_dir()));
-    }
-
-    fn path(path: Result<String, env::VarError>, default: &str) -> PathBuf {
-        Path::new(&path.unwrap_or(default.to_string())).to_owned()
-    }
 }
